refactor(packages,outputters): Move pdf:bookmark and pdf:metadata logic to the outputter
