+++ outputters/base.lua
@@ -40,6 +40,10 @@ function outputter.enterLinkTarget (_, _, _) end
 
 function outputter.leaveLinkTarget (_, _, _, _, _, _, _) end
 
+function outputter.setMetadata (_, _, _) end
+
+function outputter.setBookmark (_, _, _) end
+
 function outputter:getOutputFilename ()
   local fname
   if SILE.outputFilename then
